+++ test/unit/fuzz-tools.test.js
@@ -95,6 +95,7 @@ import * as nephroV92 from '../../lib/nephro-v92.js';
 import * as hepgiV93 from '../../lib/hepgi-v93.js';
 import * as hemoncV94 from '../../lib/hemonc-v94.js';
 import * as neuroV95 from '../../lib/neuro-v95.js';
+import * as psychV96 from '../../lib/psych-v96.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -134,6 +135,7 @@ const MODULES = {
   'hepgi-v93.js': hepgiV93,
   'hemonc-v94.js': hemoncV94,
   'neuro-v95.js': neuroV95,
+  'psych-v96.js': psychV96,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];